+++ NEWS.md
@@ -1,6 +1,8 @@
 # NMsim 0.1.1
 
 ## New features
+* Provide parameter table for simulations
+
 * seeds
 
 ## Bugfixes